refactor: 임시 이미지 파일 삭제 로직 추가
임시 이미지 파일 삭제 로직 추가

+++ src/tmpImg/TmpImg.repository.ts
@@ -1,6 +1,7 @@
 import { InternalServerErrorException } from '@nestjs/common';
 import { TmpImageEntity } from 'src/entity/tmpImg.entity';
 import { EntityRepository, Repository } from 'typeorm';
+import * as fs from 'fs';
 
 @EntityRepository(TmpImageEntity)
 export class TmpImgRepository extends Repository<TmpImageEntity> {
@@ -24,17 +25,33 @@ export class TmpImgRepository extends Repository<TmpImageEntity> {
    *@description 기간이 만료된 tmp 이미지를 모두 제거
    */
   async deleteAllTmpImg() {
+    // 한국 시간 설정 값
     const KOREAN_TIME_THRES_VALUE = 18 * 60 * 60 * 1000;
     const currentTime = new Date();
     const utc =
       currentTime.getTime() + currentTime.getTimezoneOffset() * 60 * 1000;
 
-    const DIFF_DELETE_THRES = 5 * 60 * 1000;
+    // 임시 이미지가 생성하고 유지될 시간
+    const DIFF_DELETE_THRES = 180 * 60 * 1000;
 
+    // 삭제 기준 시간
     const deleteTime = new Date(
       utc + KOREAN_TIME_THRES_VALUE - DIFF_DELETE_THRES,
     );
 
+    const res = await this.createQueryBuilder()
+      .select()
+      .where('createdAt < :deleteTime', { deleteTime })
+      .getMany();
+
+    if (!res.length) return;
+
+    // 임시 파일 삭제 로직
+    res.forEach(async (item) => {
+      fs.unlink(`${item.filename}`, (err) => {});
+    });
+
+    // db에 임시 파일 데이터 삭제
     this.createQueryBuilder()
       .delete()
       .where('createdAt < :deleteTime', { deleteTime })
